dynamically rewrite collided versions
