+++ package.json
@@ -27,6 +27,7 @@
     "@antora/cli": "^3.1.1",
     "@antora/site-generator": "^3.1.1",
     "@jcahills/antora-link-checker": "^1.0.1",
+    "@djencks/antora-aggregate-collector": "^0.1.0-beta.1",
     "asciidoctor-kroki": "^0.15.0",
     "ngrok": "^4.2.2",
     "serve": "^13.0.2"
